requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-genai>=1.0.0
pyttsx3>=2.90
edge-tts>=6.1.0
Pillow>=10.0.0
